requirements.txt
fastapi==0.155.0
uvicorn==0.30.6
pydantic==2.9.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
shap==0.46.0
lime==0.2.0.1
joblib==1.4.2
requests==2.32.3
pyyaml==6.0.2

# training addons
opensearch-py==2.7.1
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
shap==0.46.0
lime==0.2.0.1
pyyaml==6.0.2
drain3==0.9.11
